+++ app/src/main/java/com/umc/upstyle/data/model/ClothKind.kt
@@ -1,7 +1,7 @@
 package com.umc.upstyle.data.model
 
 data class ClothKind(
-    val kindId: Long,
+    val kindId: Int,
     val kindName: String,
     val thumbnailUrl: String
 )
