+++ app/src/main/java/com/umc/upstyle/data/model/ClothesCategoryResponse.kt
@@ -22,21 +22,21 @@ data class ClothesCategoryResponse(
 
     @Parcelize
     data class ClothPreview(
-        val id: Long,
-        val kindId: Long,
+        val id: Int,
+        val kindId: Int,
         val kindName: String,
-        val categoryId: Long,
+        val categoryId: Int,
         val categoryName: String,
-        val fitId: Long,
+        val fitId: Int,
         val fitName: String,
-        val colorId: Long,
+        val colorId: Int,
         val colorName: String,
         val additionalInfo: String?,
         val ootd: Ootd?
     ) : Parcelable {
         @Parcelize
         data class Ootd(
-            val id: Long,
+            val id: Int,
             val date: String,
             val imageUrls: List<String>
         ) : Parcelable
